refactor: pr review updates
