+++ web/routes.go
@@ -38,7 +38,7 @@ func getRoutes() *mux.Router {
 	router.HandleFunc("/typemap", getTypeMap).Methods("GET")
 	router.HandleFunc("/filepaths", getSchemaAndReportFile).Methods("GET")
 	router.HandleFunc("/typemap/global", setTypeMapGlobal).Methods("POST")
-	router.HandleFunc("/typemap/table", setTypeMapTableLevel).Methods("POST")
+	router.HandleFunc("/typemap/table", updateTableSchema).Methods("POST")
 	router.HandleFunc("/checkinterleave/table", checkForInterleavedTables).Methods("GET")
 	return router
 }
